refactor: split day check-in into energy and time steps and standardize scroll physics across planning screens (#ritual-82)
